Refactor shell component <Body /> sub-component into a more description <Content /> component
