--- packages/dotcom-ui-shell/src/__test__/components/Body.test.tsx
@@ -1,28 +1,0 @@
-import React from 'react'
-import renderer from 'react-test-renderer'
-import Subject from '../../components/Body'
-
-describe('dotcom-ui-shell/src/components/Body', () => {
-  describe('with stringified contents', () => {
-    it('renders the given HTML string without escaping', () => {
-      const contents = '<p>Hello, World!</p>'
-      const tree = renderer.create(<Subject contents={contents} />).toJSON()
-
-      expect(tree).toMatchSnapshot()
-    })
-  })
-
-  describe('with component contents', () => {
-    const contents = <p>Hello, World!</p>
-    const tree = renderer.create(<Subject contents={contents} />).toJSON()
-
-    expect(tree).toMatchSnapshot()
-  })
-
-  describe('with custom attributes', () => {
-    const attributes = { 'data-foo': 'bar', 'data-baz': 'qux' }
-    const tree = renderer.create(<Subject contents="" {...attributes} />).toJSON()
-
-    expect(tree).toMatchSnapshot()
-  })
-})
